feat: move columns files from .js to .ts
