--- src/resources/columns/column.model.js
@@ -1,15 +1,0 @@
-const { v4: uuidv4 } = require('uuid');
-
-class Column {
-  constructor({
-    id = uuidv4(),
-    title = 'COLUMN',
-    order = 0,
-  } = {}) {
-    this.id = id;
-    this.title = title;
-    this.order = order;
-  }
-}
-
-module.exports = Column;
